[#278] Move handle(Sender|Receiver)Open methods to base class.
Eliminate identical implementations that are part of HonoMessaging and
SimpleDeviceRegistrationService. Splitted test classes.
