+++ services/messaging/src/test/java/org/eclipse/hono/messaging/HonoMessagingTest.java
@@ -13,39 +13,22 @@ package org.eclipse.hono.messaging;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
-import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.*;
 
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
-import org.apache.qpid.proton.amqp.transport.Target;
 import org.apache.qpid.proton.engine.Record;
 import org.apache.qpid.proton.engine.impl.RecordImpl;
-import org.apache.qpid.proton.message.Message;
-import org.eclipse.hono.auth.Activity;
 import org.eclipse.hono.auth.HonoUser;
-import org.eclipse.hono.config.ServiceConfigProperties;
-import org.eclipse.hono.messaging.HonoMessaging;
-import org.eclipse.hono.messaging.HonoMessagingConfigProperties;
-import org.eclipse.hono.service.amqp.BaseEndpoint;
 import org.eclipse.hono.service.amqp.Endpoint;
-import org.eclipse.hono.service.auth.AuthorizationService;
-import org.eclipse.hono.telemetry.TelemetryConstants;
 import org.eclipse.hono.util.Constants;
-import org.eclipse.hono.util.ResourceIdentifier;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 
 import io.vertx.core.Context;
-import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.eventbus.EventBus;
 import io.vertx.proton.ProtonConnection;
-import io.vertx.proton.ProtonReceiver;
 
 /**
  * Unit tests for Messaging Service.
@@ -80,73 +63,6 @@ public class HonoMessagingTest {
     }
 
     @Test
-    public void testHandleReceiverOpenForwardsToTelemetryEndpoint() throws InterruptedException {
-
-        // GIVEN a server with a telemetry endpoint
-        final ResourceIdentifier targetAddress = ResourceIdentifier.from(TelemetryConstants.TELEMETRY_ENDPOINT, Constants.DEFAULT_TENANT, null);
-        final CountDownLatch linkEstablished = new CountDownLatch(1);
-        final Endpoint telemetryEndpoint = new BaseEndpoint<ServiceConfigProperties>(vertx) {
-
-            @Override
-            public String getName() {
-                return TelemetryConstants.TELEMETRY_ENDPOINT;
-            }
-
-            @Override
-            public void onLinkAttach(final ProtonConnection con, final ProtonReceiver receiver, final ResourceIdentifier targetResource) {
-                linkEstablished.countDown();
-            }
-
-            @Override
-            protected boolean passesFormalVerification(ResourceIdentifier targetAddress, Message message) {
-                return true;
-            }
-        };
-        AuthorizationService authService = mock(AuthorizationService.class);
-        when(authService.isAuthorized(Constants.PRINCIPAL_ANONYMOUS, targetAddress, Activity.WRITE)).thenReturn(Future.succeededFuture(Boolean.TRUE));
-        HonoMessaging server = createServer(telemetryEndpoint);
-        server.setAuthorizationService(authService);
-
-        // WHEN a client connects to the server using a telemetry address
-        final Target target = getTarget(targetAddress);
-        final ProtonReceiver receiver = mock(ProtonReceiver.class);
-        when(receiver.getRemoteTarget()).thenReturn(target);
-        when(receiver.attachments()).thenReturn(mock(Record.class));
-        server.handleReceiverOpen(newConnection(Constants.PRINCIPAL_ANONYMOUS), receiver);
-
-        // THEN the server delegates link establishment to the telemetry endpoint 
-        assertTrue(linkEstablished.await(1, TimeUnit.SECONDS));
-    }
-
-    @Test
-    public void testHandleReceiverOpenRejectsUnauthorizedClient() throws InterruptedException {
-
-        // GIVEN a server with a telemetry endpoint
-        final ResourceIdentifier restrictedTargetAddress = ResourceIdentifier.from(TelemetryConstants.TELEMETRY_ENDPOINT, "RESTRICTED_TENANT", null);
-        final Endpoint telemetryEndpoint = mock(Endpoint.class);
-        when(telemetryEndpoint.getName()).thenReturn(TelemetryConstants.TELEMETRY_ENDPOINT);
-        AuthorizationService authService = mock(AuthorizationService.class);
-        when(authService.isAuthorized(Constants.PRINCIPAL_ANONYMOUS, restrictedTargetAddress, Activity.WRITE)).thenReturn(Future.succeededFuture(Boolean.FALSE));
-        HonoMessaging server = createServer(telemetryEndpoint);
-        server.setAuthorizationService(authService);
-
-        // WHEN a client connects to the server using a telemetry address for a tenant it is not authorized to write to
-        final CountDownLatch linkClosed = new CountDownLatch(1);
-        final Target target = getTarget(restrictedTargetAddress);
-        final ProtonReceiver receiver = mock(ProtonReceiver.class);
-        when(receiver.getRemoteTarget()).thenReturn(target);
-        when(receiver.setCondition(any())).thenReturn(receiver);
-        when(receiver.close()).thenAnswer(invocation -> {
-            linkClosed.countDown();
-            return receiver;
-        });
-        server.handleReceiverOpen(newConnection(Constants.PRINCIPAL_ANONYMOUS), receiver);
-
-        // THEN the server closes the link with the client
-        assertTrue(linkClosed.await(1, TimeUnit.SECONDS));
-    }
-
-    @Test
     @SuppressWarnings({"rawtypes", "unchecked"})
     public void testServerPublishesConnectionIdOnClientDisconnect() {
 
@@ -174,13 +90,6 @@ public class HonoMessagingTest {
         HonoMessaging server = createServer(null);
         assertThat(server.getPortDefaultValue(), is(Constants.PORT_AMQPS));
         assertThat(server.getInsecurePortDefaultValue(), is(Constants.PORT_AMQP));
-    }
-
-
-    private static Target getTarget(final ResourceIdentifier targetAddress) {
-        Target result = mock(Target.class);
-        when(result.getAddress()).thenReturn(targetAddress.toString());
-        return result;
     }
 
     private static ProtonConnection newConnection(final HonoUser user) {
